nxp-s32g2xx: pkcs11-hse: update to bsp31 rc3
Issue: LINUXEXEC-16831

This patch will update pkcs11-hse recipe to compatible with NXP SDK bsp31
rc3. In this update, the pkcs demo has been restructured into two libs:
	libpkcs-hse.so and libhse.so
which will interact with the linux uio-hse driver to support the hse demo.
More over, due to the rc3 update, the bb file is modified accordingly and
add a patch:
	0001-pkcs11-hse-Makefile-using-internal-compile-variables.patch
to use bitbake internal variable CC to replace the CROSS_COMPILE variable
in Makefile of source codes, or else the software will build fail with the
CROSS_COMPILE not set.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>