nxp-s32g2xx: linux: update to bsp31 rc3
Issue: LINUXEXEC-16831

This patch will update linux-yocto recipe to compatible with NXP SDK bsp31
rc3. And for hse driver splite to uio driver and crypto driver, so drop the
unneeded patches for hse demo. To be more clear, there are 20 kernel patches
for rc3, but the 0009, 0010, 0011, and 0012 four patches are all defconfig
related, we dont need them so dropped them.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>