nxp-s32g2xx: llce: update to compatible with S32G3 RDB3
Issue: LINUXEXEC-19151

Given llce firmwares are different between S32G2 and S32G3 boards, so the
recipe needs to update, and the update infos as following:

  1. create a new llce.inc file to include the common settings of llce,
     so these infos can be shared in the whole bsp layer.
  2. use pkg_postinst_ontarget() to create symlinks for llce firmwares
     based on the board type(S32G3 or S32G2).
  3. simplify the install and deploy codes for llce.
  4. update README file to compatible with llce recipe.

And given the dependency of llce kos, there are two loops for rmmod and
modprobe operations in the pkg_postinst_ontarget(). In detail, the module
llce_mailbox is depended by llce_can/llce_core, so it should be removed at
last, if use one loop, the rmmod of llce_mailbox will fail. And the two
loops could make the re-installation have a clear environment.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>