nxp-s32g: ipc-shm: update to support both s32g2 and s32g3 platforms
Issue: LINUXEXEC-19151

Now the RDB3 has been supported, and the ipcf recipe need to update to
support both s32g2 and s32g3 platforms. In detail, the ipcf source codes
will use variable PLATFORM_FLAVOR to check the platform, so the function
do_compile() is overwritten and the related kernel modules are renamed
based on s32g2 or s32g3.

Moreover, add pkg_postinst_ontarget() to delete the unneeded ipcf modules
for the runtime platform, if not, modprobe will fail due to modules
dependency. And the README is updated too.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>