nxp-s32g: pfe: update to fix dma share memory init failed issue
Issue: LINCD-9031

This patch is to fix the pfe dma share memory init failed issue, which is
caused by the kernel commit fd3e865fb9d0 ("dts: fsl-s32g: pfe: Add
reserved-memory node for the routing table"), this commit adds one reserved
memory node for routing table of pfe, but did not add driver codes to
compatible with the dts update.

To fix the failure, there added a function to check the routing table
reserved memory node status, then the idx will be correct to the right
value and the init will pass.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>