nxp-s32g: linux: add patches to fix pfe build and probe failure
Issue: LINCD-9031

The pfe resvered memory nodes in dts file are updated in bsp33 rc1, but
the pfe driver probe failed due to the updated reserved memory nodes.
This patch is to fix it.

Moreover, this patch will revert the commit 1d6cd3929360 ("modpost: turn
missing MODULE_LICENSE() into error"), to fix the pfe build error due to
lack of MODULE_LICENSE in the pfe source codes. And it will be dropped
when NXP fixed the pfe MODULE_LICENSE issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>