nxp-s32g: support for verifying fitImage with the public key stored in atf dtb file
Issue: LINUXEXEC-20369

Along with the NXP s32g upgrads continusly, the device tree in u-boot
is moved into atf. And BL33 is pointed to u-boot binary file without
device tree. Based on this modification, public key needs to be saved
in atf dtb instead of being saved in u-boot dtb file. So, add code to
support for verifying fitImage with the public key stored in atf dtb file.

Signed-off-by: Meng Li <Meng.Li@windriver.com>