nxp-s32g: libfci: add libfci/libfci-cli recipes
Issue: LINUXEXEC-21797

The FCI(Fast Control Interface) provides additional features management for
pfeng. The libfci is the recipe for FCI library, which exposes APIs of FCI
for userspace, and the libfci-cli is the recipe for FCI userspace demo tool.

Then, with these two components, users can control or manage the pfeng ports,
in userspace. The detail informations about FCI, users can refer to the
"FCI API Reference" document from NXP. And the libfci/libfci-cli bb files
refered to the related recipes of NXP SDK bsp33 release.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>