nxp-s32g: atf: remove S32_HAS_HV setting to fix kernel complain
Issue: LIN1022-1253

The EXTRA_OEMAKE flag S32_HAS_HV is to control the running EL level of
uboot when booted by atf. By default, it is set to 0, i.e. the uboot will
boot from EL1, and this is the normal setting for S32G platform. While if
set it to 1, the uboot will boot from EL2, then config XEN_SUPPORT in uboot
should be enabled too. And in S32G settings, XEN_SUPPORT is disabled
by default, because it is not supported in wrl nxp-s32g bsp yet, thus, the
linux kernel will complain when the system starts up, as following:

  CPU: CPUs started in inconsistent modes
  WARNING: CPU: 0 PID: 1 at arch/arm64/kernel/smp.c:426 smp_cpus_done+0x78/0xa0
  Modules linked in:
  CPU: 0 PID: 1 Comm: swapper/0 Not tainted 5.15.58-rt48-yocto-preempt-rt #1
  Hardware name: NXP S32G399A-RDB3 (DT)
  pstate: 60000005 (nZCv daif PAN -UAO -TCO -DIT -SSBS BTYPE=-)
  pc : smp_cpus_done+0x78/0xa0
  lr : smp_cpus_done+0x78/0xa0
  sp : ffffffc00991be00
  x29: ffffffc00991be00 x28: 0000000000000000 x27: 0000000000000000
  x26: 0000000000000000 x25: 0000000000000000 x24: 0000000000000000
  x23: 0000000000000000 x22: ffffffc0093bc000 x21: 0000000000000001
  x20: ffffffc008f1a2a0 x19: ffffffc009294000 x18: ffffffffffffffff
  x17: 0000000000000001 x16: 0000000000000007 x15: 000000cf17f7d436
  x14: 000358cff1946410 x13: 7365646f6d20746e x12: 65747369736e6f63
  x11: 6e69206e69206465 x10: 7472617473207355 x9 : ffffffc0080a3a64
  x8 : 636e69206e692064 x7 : ffffffc009271630 x6 : 0000000000000000
  x5 : 0000000000000001 x4 : ffffff885bf39a68 x3 : 0000000000000000
  x2 : 0000000000000000 x1 : 0000000000000000 x0 : ffffff8800830000
  Call trace:
  smp_cpus_done+0x78/0xa0
  smp_init+0x74/0x84
  kernel_init_freeable+0x13c/0x294
  kernel_init+0x30/0x130
  ret_from_fork+0x10/0x20

So remove the S32_HAS_HV setting in atf to fix the above issue and make
nxp-s32g bsp back to normal status.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>