nxp-s32g: linux: add internal patch to fix "rmmod pfeng" calltrace
Issue: LIN1022-1262

This patch is to fix the following calltrace when "rmmod pfeng":

    Call trace:
       kfree+0x74/0x360
       regcache_exit+0x38/0x80
       regmap_exit+0x28/0x15c
       xpcs_release+0x2c/0x50
       serdes_phy_release+0x48/0x60
       of_phy_put+0x4c/0x80
       devm_phy_release+0x30/0x40
       release_nodes+0x7c/0x12c
       devres_release_all+0x90/0xf0
       __device_release_driver+0x1a0/0x250
       driver_detach+0xcc/0x110
       bus_remove_driver+0x64/0xc0
       driver_unregister+0x3c/0x6c
       platform_driver_unregister+0x20/0x30
       pfeng_platform_driver_exit+0x1c/0xc1c [pfeng]
       __arm64_sys_delete_module+0x184/0x210
       invoke_syscall+0x5c/0x130
       el0_svc_common.constprop.0+0x4c/0x100
       do_el0_svc+0x50/0xb4
       el0_svc+0x28/0x80
       el0t_64_sync_handler+0xa4/0x130
       el0t_64_sync+0x1a0/0x1a4

This is a kernel patch, which should be merged into linux-yocto kernel tree,
but we are about to enter into LTS22 release stage, and the auto sync for
LTS22 kernel repos have been stopped too, so put the patch internally for
nxp-s32g bsp.

And this internal patch will be dropped when the auto sync started again.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>