nxp-s32g: linux: fix sleeping function called from invalid context issue
Issue: LIN1022-1267

This series of patches are kernel patches, which should be merged into
linux-yocto kernel tree, but we are about to enter into LTS22 release stage,
and the auto sync for LTS22 kernel repos have been stopped too, so put the
patch internally for nxp-s32g bsp.

And this internal patch will be dropped when the auto sync started again.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>