nxp-s32g: atf-s32g: fix the OPENSSL_DIR assignment
Issue: LIN1022-2000

When compiling atf, the flag OPENSSL_DIR is used to provide the
installed openssl directory, and if not set, atf will use the default
value "/usr".

But this raises some problems. When the version of gcc in HostOS is under
7.5.0, WRLinux will use extended buildtools and the libs and headers of
gcc will switch to the path below:
    bin/buildtools.buildtools-extended-standalone-10.21.44.0/sysroots/x86_64-wrlinuxsdk-linux/

So we need pass the correct openssl directory path to atf, using
STAGING_DIR_NATIVE as openssl directory can fix it.

This is to fix the compiling error as below:

    In file included from fiptool.c:12:
    /usr/include/limits.h:124:3: error: #include_next is a GCC extension [-Werror]
    124 | # include_next <limits.h>

    ^~~~~~~~~~~~
    In file included from /usr/include/limits.h:124,
    from fiptool.c:12:

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>