nxp-s32g: pfe: update to align pfeng-salve with kernel codes
Issue: LINUXEXEC-23697

This patch is to update pfeng-slave driver to keep align with kernel codes.
And new dtbs for pfe master/slave mode are added too, moreover, the new
variables DTB_FILES and FIT_IMAGES are added to make the it more clear.

Besides, the unneeded patch for pfe slave driver:
  0001-pfeng-fix-to-compatible-with-kernel-v5.15.patch

is also dropped.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>