nxp-s32g: pfe: share the pfe common contents for pfe drivers and libfci components
Issue: LINUXEXEC-24056

This patch is to share the common contents of pfe for related drivers and
libfci components. Because the libfci is a userspace application of pfe, and
its codes are all included in the pfe repo too. So move the libfci bb files
to pfe recipe will simplify the bb files.

Besides, move PFE_LOCAL_FIRMWARE_DIR variable to pfe_common.inc, then
pfe-slave driver can access the pfe firmwares too.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>