nxp-s32g: atf: update to compatible with SDK BSP35 RC7
Issue: LINUXEXEC-24056

This patch is to update atf recipe to compatible with SDK BSP35 RC7 release.
But there is a boot failure issue if the following patch applied:
  0026-secboot-add-support-for-userspace-secboot.patch

So this patch is introduced but not applied for now, and it will be applied
later when the issue is fixed by vendor.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>