nxp-s32g: pkcs11-hse: update to use proper interface of HSE firmware for S32G2 and S32G3 platforms
Issue: LINUXEXEC-24056

For pkcs11-hse recipe, there are some new demo applications added recently,
like hse-secboot, hse-sysimg, hse-encrypt, and so on. And for these demos,
they depends on different interfaces of HSE firmware for S32G2 and S32G3
platforms when they are compiled. So update the pkcs11-hse source codes and
bsp recipe to add compatibility for this issue.

In detail, there are following updates:
  1. for pkcs11-hse source codes, added two HSE firmware interface
     directories: s32g2 and s32g3, in hse-fw directory
  2. for pkcs11-hse recipe, use the right HSE_FWDIR location when compile
     the codes, for different platforms. And use pkg_postinst_ontarget()
     script to put the right demo binaries to /usr/bin, when the rootfs
     first starts up.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>