nxp-s32g: simplify the building process of generating fitImage
Issue: LINUXEXEC-24285

Based on Current building process, there will be 4 fitImage images with
respective suffix name for 4 boards. But this is redundant process, because
the fitImage image are the same with each other, only with suffix name.
Since only one fitImage is able to support for 4 board, it doesn’t need to
generate multiple image. Therefore, remove the useless for loop to simplify
the building process of generating fitImage
In additional, modify relative configurations and description in README.

Signed-off-by: Meng Li <Meng.Li@windriver.com>