nxp-s32g: adjust the kernel load and entry address of fitImage
Issue: LIN1022-3263, LINCD-11657

When booting up fitImage, there is a warning info as below:
[Firmware Bug]: Kernel image misaligned at boot, please fix your
bootloader!
This warning info is introduced by upstream commit("arm64: warn
on incorrect placement of the kernel by the bootloader"). It requests
that the entry address of kernel should align with 2MB at least.
So, adjust the kernel load and entry address to compatible with the
upsreatm commit.

Signed-off-by: Meng Li <Meng.Li@windriver.com>