nxp-s32g: conf: add qspi boot support for S32G platforms
Issue: LINUXEXEC-25381

This patch is mainly to add qspi uboot defconfig in conf for S32G platforms,
and the variables XXX_UBOOT_DEFCONFIG_NAME are not needed anymore, so drop
them.

And make_nxp_s32g_wic.inc is updated too to get the correct plat names for
WIC images, which are only for SD/EMMC boot mode.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>