nxp-s32g: improve bootloader for parallel secure boot feature
Issue: LINUXEXEC-25112

In parallel secure boot mode, hse firmware brings up m7 code and fip.bin
separately. But at the first time of booting up, it is in non-secure boot
mode, and it needs to bring up fip.bin directly because m7 code doesn't
bring up a53 any more. In this way, secure boot is able to be enabled with
u-boot command.
In addtional, save m7 entry address in reserve space, and it will be
used to config m7 boot in u-boot command.

Signed-off-by: Meng Li <Meng.Li@windriver.com>