nxp-s32g: linux: update to compatible with SDK BSP36 rc5
Issue: LINUXEXEC-25380

The BSP36 rc5 has been released, and this patch is to update linux recipe to
compatible with it.

And there are 3 new added internal patches as following:

  0001-s32cc-update-to-align-with-bsp36-v5.15-kernel.patch
  0002-xpcs-s32-rename-fsl-s32gen1-xpcs.h-to-nxp-s32cc-xpcs.patch
  0003-xpcs-s32-rename-functions-s32gen1_xxx-to-s32cc_xxx-i.patch

to make wrlinux s32g kernel(based on SDK v5.10 kernel) more compatible with
SDK v5.15 kernel.

And there is a nxp-s32g.cfg file to add related kernel configs to align with
SDK.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>