nxp-s32g: force to use the printf in /usr/bin directory
Issue: LINUXEXEC-26092

When using the dash as the default shell, the str2bin() is not able
to generate the expected binary data. This is because that the printf
builtin in dash doesn't support the '\xHH' escape sequence. So add
the absolute path to make sure that the standalone printf is used.

Signed-off-by: Meng Li <Meng.Li@windriver.com>