nxp-s32g: pkcs11-hse: add hse firmware version variables
Issue: LINUXEXEC-26603

To make it more convenient to use and maintain pkcs11-hse recipe,
the following two variables are added to specify the HSE firmware versions
for S32G2 and S32G3 respectively, then users can set them to the preferred
versions to use.

And the tarball of pkcs11-hse is not needed anymore, so delete it.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>