pkcs11-hse: Add support for the read only rootfs
Issue: LINUXEXEC-27973

The pkg_postinst_ontarget method is used by this recipe, it will emit
build error if we enable the read-only-rootfs image feature. In order
to fix this issue, introduce a S32G_SOC_VARIANT and leverage it to
pick up the corresponding pkcs11-hse examples when read-only-rootfs is
enabled.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>