nxp-s32g: set load_fitimage_addr if if secure boot layer is enabled
Issue: LIN1023-962

When building nxp-s32g project with secure boot layer, the fitImage is
created and deployed in the boot partition. So, it needs to bring up the
fitImage with bootm command. Set load_fitimage_addr so that the correct
bringing up command is used in meta-lat/recipes-bsp/u-boot/u-boot-uenv_2.0.bb.

Signed-off-by: Meng Li <Meng.Li@windriver.com>