nxp-s32g: atf: update to compatible with SDK BSP38 release
Issue: LINUXEXEC-29546

The SDK BSP38 has been released, this patch is to update atf recipe to
compatible with it.

And the related internal patches have been updated accordingly.
Moreover, the patch "0001-s32-extend-the-DTB-size-for-BL33.patch" is moved
into the atf-s32g_2.5.bb from s32g-secure-boot's atf-s32g_2.5.bbappend
to make it a comman patch for atf-s32g.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>