nxp-s32g: uboot: update to compatible with SDK BSP38 release
Issue: LINUXEXEC-29546

The SDK BSP38 has been released, this patch is to update uboot recipe to
compatible with it.

And the related internal patches have been updated accordingly.
Moreover, the patch "0001-s32-extend-the-DTB-size-for-BL33.patch" is moved
into u-boot-s32.inc from s32g-secure-boot's u-boot-s32_2020.04.bbappend to
make it a comman patch for u-boot-s32.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>