nxp-s32g: README: update to make it more flexible and compatible with latest bsp status
Issue: LIN1023-3321

The nxp-s32g bsp layer has been updated to SDK BSP38 for a period of time,
so update README to compatible with it.

Moreover, since the S32G SDK is under development, the reference documents
of SDK and the related firmwares will updated in every SDK release, so the
related description of SDK user manual and firmware info are changed
accordingly to make it more convenient and flexible to maintain and read.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>