nxp-s32g: install the final bootloader image into boot directory
Issue: LINUXEXEC-31031

When secure boot is enabled, the atf-<platform>.s32 image is
also modified in do_deploy. So, if the bootloader image is only
copied into boot directory in do_install, it is not the final one.
therefore, it also needs to be copied in do_deploy.
In addition, when m7 boot feature is enable, the m7 bootloaer
Image also needs to be copied into boot directory both secure
and non-secure boot mode.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>