nxp-s32g: improve the secure boot building architecture
Issue: LINUXEXEC-31031

In commit 9eeeba3e81c8("nxp-s32g: install the final bootloader image
into boot directory "), move the do_deploy before do_package so that
the final bootloader is installed into boot directory when secure boot
feature is enabled. But in this way, if dm-verity feature is enabled,
there will be circular dependency reported. This issue is introduced
by interdependence of atf and kernel. Kernel depends on do_populate_sysroot
of atf, and atf depends on do_deploy of kernel. In order to aoivd the
interdependence between kernel and atf. Refer to newer oe-core code,
remove the dependency of atf from kernel recipe. Kernel also can be
built separately even if secure boot is enabled. fitimage_assemble only
signs fitImage, and doesn't need to add the public key into atf dtb file.
The action of adding public key into atf dtb is implemented in atf building
stage. Based on above improvement, the interdependence of atf and kernel
is removed so that secure boot and dm-verify are able to be enabled at
the same time.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>