nxp-s32g: uboot: fix EMMc boot failure of RDB2 board
Issue: LIN1023-5077

This patch reverts the commit b5a0c8f587f7 ("s32cc: mmc: Enable
'MMC_IO_VOLTAGE'"), which is introduced by SDK BSP40. But this
commit will cause the EMMc boot failure on our RDB2 (REV.C) boards.
The reason is that, the voltage supply of eMMC on RDB2 REV.C in default
is 3.3V, but the default mode of BSP40 is using 1.8v, so the EMMc boot
failed.

So revert the commit to fix the issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>