nxp-s32g: pfe: update to fix pfe1 2.5G up/down failure issue
Issue: LIN1023-5412

On RDB3 board, the default emac1 is connected with AQR113C phy, and the
supported speed is up to 2.5G(SGMII enaled). But when it runs in 2.5G,
if we bring the PFE down and then bring it back up, we will encounter the
following error:

  # ifconfig pfe1 down
  pfeng 46000000.pfe pfe1: Link is Down
  # ifconfig pfe1 up
  pfeng 46000000.pfe pfe1: validation of 2500base-x with support 00000000,00018000,000e706c and advertisement 00000000,00008000,0000606c failed: -EINVAL
  pfeng 46000000.pfe pfe1: ERR: (DRIVER) event 1 - Driver runtime error: [pfeng-phylink.c:386] could not attach PHY: -22
  pfeng 46000000.pfe pfe1: ERR: (DRIVER) event 1 - Driver runtime error: [pfeng-netif.c:147] Error connecting to the phy: -22
  SIOCSIFFLAGS: Invalid argument

The reason is that, when pfe1 is running in 2.5G, the related phy device
will set the interface mode to PHY_INTERFACE_MODE_2500BASEX, which is not
valid in pfeng_phylink_validate(), and if we try to add the above
PHY_INTERFACE_MODE_2500BASEX to the valid list, like SGMII and R/G/MII modes,
the default speed of pfe1 will be 1G instead, which is not reasonable.

So let's set it to PHY_INTERFACE_MODE_SGMII just in one own code block,
to fix the issue and reduce the other side effects.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>