nxp-s32g: secure boot: remove the unneeded do_generate_secboot_script() in atf
Issue: LINUXEXEC-33843

The do_generate_secboot_script() is introduced when adding the
atf-hse-secboot.inc file. It aims to generate a script for secboot
initialisation, making it simple for users, but in wrlinux case,
the --bl2_bin, --bl2_sign etc. params should be set with the specific
board's SOC information. And given the secboot initialisation is a
one-time command, the do_generate_secboot_script() is not so required,
so remove it.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>