nxp-s32g: u-boot-tools: add anon function to compile the native recipe only for S32G platform
Issue: 	LIN1024-8758

The recipes u-boot-tools_2022.04-s32.bb provides a native package only for
S32G platform.

In this recipe, the COMPATIBLE_MACHINE is used to make it only for S32G
SOCs. But for native package, the COMPATIBLE_MACHINE actually can not make
it to only be compiled for S32G platform bsps, so a new anon python function
is introduced to check the MACHINE variable to decide if the recipe should
be compiled or not.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>